+++ config/sample.exs
@@ -22,12 +22,21 @@ config :passport, Passport.DataStore.EctoAdapter,
   identity_field:             :username,
   password_field:             :password,
   email_field:                :email,
-  password_reset_token_field: :password_reset_token
+  password_reset_token_field: :password_reset_token,
+  email_confirmed_field:      :email_confirmed
 
-# Secret key to use for Password Reset JWT signing (REQUIRED)
+# Secret key to use for signing password reset JWTs (Required if using)
 config :passport, Passport.ResetPassword,
     key: "USE_ENV_VARIABLE"
 
 # Reset password options
 config :passport, Passport.ResetPassword,
-    timeout: 60 * 60 * 2
+  timeout: 60 * 60 * 2
+
+# Secret key to use for signing the email confirmation JWTs (Required if using)
+config :passport, Passport.ConfirmEmail,
+  key: "USE_ENV_VARIABLE"
+
+# Confirm email options
+config :passport, Passport.ResetPassword,
+  timeout: 60 * 60 * 48